+++ demo-app/src/androidTest/java/cash/z/wallet/sdk/sample/demoapp/SampleCodeTest.kt
@@ -214,6 +214,7 @@ class SampleCodeTest {
                     // Using existing wallet init mode as simplification for the test
                     walletInitMode = WalletInitMode.ExistingWallet,
                     zcashNetwork = network,
+                    isTorEnabled = false
                 )
             }