+++ sdk-lib/src/androidTest/java/cash/z/ecc/android/sdk/util/BalancePrinterUtil.kt
@@ -101,6 +101,7 @@ class BalancePrinterUtil {
                             // Using existing wallet init mode as simplification for the test
                             walletInitMode = WalletInitMode.ExistingWallet,
                             zcashNetwork = network,
+                            isTorEnabled = false,
                         )
 
 //            deleteDb(dataDbPath)