+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/FastestServerFetcher.kt
@@ -3,6 +3,7 @@ package cash.z.ecc.android.sdk.internal
 import cash.z.ecc.android.sdk.internal.model.ext.toBlockHeight
 import cash.z.ecc.android.sdk.model.BlockHeight
 import cash.z.ecc.android.sdk.model.FastestServersResult
+import cash.z.ecc.android.sdk.model.SdkFlags
 import cash.z.ecc.android.sdk.model.ZcashNetwork
 import cash.z.ecc.android.sdk.util.WalletClientFactory
 import co.electriccoin.lightwallet.client.CombinedWalletClient
@@ -35,6 +36,7 @@ internal class FastestServerFetcher(
     private val backend: TypesafeBackend,
     private val network: ZcashNetwork,
     private val walletClientFactory: WalletClientFactory,
+    private val sdkFlags: SdkFlags
 ) {
     operator fun invoke(servers: List<LightWalletEndpoint>): Flow<FastestServersResult> =
         flow {
@@ -118,12 +120,20 @@ internal class FastestServerFetcher(
         val remoteInfo: LightWalletEndpointInfoUnsafe?
         val getServerInfoDuration =
             measureTime {
-                // TODO [#1772]: redirect to correct service mode after 2.1 release
                 // 5 seconds timeout in case server is very unresponsive
                 remoteInfo =
                     withTimeoutOrNull(5.seconds) {
                         when (
-                            val response = lightWalletClient.getServerInfo(ServiceMode.Direct)
+                            val response =
+                                lightWalletClient.getServerInfo(
+                                    if (sdkFlags.isTorEnabled) {
+                                        ServiceMode.Group(
+                                            "validateServerEndpointAndMeasure(${endpoint.host}:${endpoint.port})"
+                                        )
+                                    } else {
+                                        ServiceMode.Direct
+                                    }
+                                )
                         ) {
                             is Response.Success -> response.result
                             is Response.Failure -> {
@@ -163,12 +173,18 @@ internal class FastestServerFetcher(
         val currentChainTip: BlockHeight
         val getLatestBlockHeightDuration =
             measureTime {
-                // TODO [#1772]: redirect to correct service mode after 2.1 release
                 currentChainTip =
                     when (
                         val response =
                             lightWalletClient.getLatestBlockHeight(
-                                serviceMode = ServiceMode.Direct
+                                serviceMode =
+                                    if (sdkFlags.isTorEnabled) {
+                                        ServiceMode.Group(
+                                            "validateServerEndpointAndMeasure(${endpoint.host}:${endpoint.port})"
+                                        )
+                                    } else {
+                                        ServiceMode.Direct
+                                    }
                             )
                     ) {
                         is Response.Success -> {