+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/model/SdkFlags.kt
@@ -1,0 +1,5 @@
+package cash.z.ecc.android.sdk.model
+
+data class SdkFlags(
+    val isTorEnabled: Boolean
+)